server/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
beanie==1.28.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cramjam==2.9.1
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fonttools==4.55.3
fpdf2==2.8.2
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
idna==3.10
kiwisolver==1.4.8
lazy-model==0.2.0
lxml==5.3.0
matplotlib==3.10.0
motor==3.6.0
multitasking==0.0.11
numpy==2.2.1
packaging==24.2
pandas==2.2.3
passlib==1.7.4
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.9.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-snappy==0.7.3
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
winkerberos==0.12.0
yfinance==0.2.51
